+++ src/components/Fixture.js
@@ -1,16 +1,16 @@
-import "./Results.scss";
-import "./Card.scss";
+import "./Fixture.scss";
 import acbelfastlogo from "../images/acbelfastlogo.png";
 import southsidelogo from "../images/southsidelogo.jpeg";
+import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
 
 
-const Results = () => {
+const Fixture = () => {
   return (
-    <div className="card__container column__half column__half--gap">
-      <h2>Results</h2>
+    <div className="card__container column__half">
+      <h2>Fixtures</h2>
       <div className="card">
         <div className="results__title">
-        <b className="results__title--link">View more results</b>
+        <Link to="/fixtures"><b className="results__title--link">View more fixtures</b></Link>
         </div>
         <div className="match__details">
         <h3 className="match__dateTime">Friday 20 January 2022</h3> {' '}
@@ -19,7 +19,7 @@ const Results = () => {
         </div>
         <div className="match__teams">
         <img src={acbelfastlogo} width="75px" height="70px"></img>
-        <h4 className="match__result">AC Belfast 0 - 1 Southside II</h4>
+        <h4 className="match__result">AC Belfast v Southside II</h4>
         <img src={southsidelogo} width="75px" height="70px"></img>
         </div>
         <hr className="match__divider"/>
@@ -30,7 +30,7 @@ const Results = () => {
         </div>
         <div className="match__teams">
         <img src={acbelfastlogo} width="75px" height="70px"></img>
-        <h4 className="match__result">AC Belfast 35s 0 - 1 Southside II</h4>
+        <h4 className="match__result">AC Belfast 35s v Southside II</h4>
         <img src={southsidelogo} width="75px" height="70px"></img>
         </div>
       </div>
@@ -38,4 +38,4 @@ const Results = () => {
   );
 };
 
-export default Results;
+export default Fixture;
