+++ src/components/Fixture.scss
