+++ src/components/Header.js
@@ -1,13 +1,11 @@
 import "./Header.scss";
 import "../App.css";
-import Navbar from "./Navbar";
 import footballHero from "../images/footballhero.jpg";
 
 
 function Header() {
   return (
       <header>
-        <Navbar/>
         <div className="heading__container">
           <h1 className="hdg-founded">Founded in 2022, Athletic Club Belfast is a</h1>
           <h1 className="hdg--based">Senior Amateur Men’s Football Club with a first and an over 35s team based in Belfast, Northern Ireland. </h1>
