+++ src/components/Result.scss
@@ -46,7 +46,8 @@
 
 .match__divider {
     margin: 2rem 0;
-    border-top: 1px solid #fff;
+    border-top: 0px solid #fff;
+    width: 90%;
 
 }
