+++ src/components/Squads.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Squads = () => {
+    return (
+        <div>
+            <h3 style={{marginTop: '5rem', color: 'white'}}>Squads</h3>
+        </div>
+    )
+}
+
+export default Squads
