+++ src/components/contact/Contact.js
@@ -1,0 +1,10 @@
+
+const Contact = () => {
+    return (
+        <div>
+            <h3 style={{marginTop: '5rem', color: 'white'}}>Contact</h3>
+        </div>
+    )
+}
+
+export default Contact
