+++ src/components/fixtures/Fixtures.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Fixtures = () => {
+    return (
+        <div>
+            <h3 style={{marginTop: '5rem', color: 'white'}}>Fixtures</h3>
+        </div>
+    )
+}
+
+export default Fixtures
