+++ src/components/results/Results.js
@@ -1,0 +1,9 @@
+const Results = () => {
+  return (
+    <div>
+      <h3 style={{ marginTop: "5rem", color: "white" }}>Results</h3>
+    </div>
+  );
+};
+
+export default Results;
